database and context creation changes
