+++ src/main/java/com/skapp/SkappApplication.java
@@ -5,9 +5,6 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.scheduling.annotation.AsyncConfigurer;
 import org.springframework.scheduling.annotation.EnableAsync;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
-
-import java.util.concurrent.Executor;
 
 @EnableAsync
 @SpringBootApplication
@@ -17,17 +14,6 @@ public class SkappApplication implements AsyncConfigurer {
 
 	public static void main(String[] args) {
 		SpringApplication.run(SkappApplication.class, args);
-	}
-
-	@Override
-	public Executor getAsyncExecutor() {
-		ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
-		executor.setThreadNamePrefix("Async-Thread-");
-		executor.setCorePoolSize(10);
-		executor.setMaxPoolSize(20);
-		executor.setQueueCapacity(500);
-		executor.initialize();
-		return executor;
 	}
 
 }
