+++ src/main/java/com/skapp/community/common/component/AsyncConfiguration.java
@@ -1,0 +1,25 @@
+package com.skapp.community.common.component;
+
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
+
+import java.util.concurrent.Executor;
+
+@Configuration
+@ConditionalOnMissingBean(name = "EpAsyncConfiguration")
+public class AsyncConfiguration {
+
+	@Bean
+	public Executor getAsyncExecutor() {
+		ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
+		executor.setThreadNamePrefix("Async-Thread-");
+		executor.setCorePoolSize(10);
+		executor.setMaxPoolSize(20);
+		executor.setQueueCapacity(500);
+		executor.initialize();
+		return executor;
+	}
+
+}
